requirements.txt
python-dotenv>=1.0.0
youtube-dl>=2021.12.17
google-api-python-client>=2.86.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
openai-whisper>=20231117
google-generativeai>=0.3.1
pydub>=0.25.1
tqdm>=4.65.0